+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.85.5] - 2026-09-05
 
 ### Added